app/requirements.txt
streamlit
psycopg2-binary
torch --index-url https://download.pytorch.org/whl/cpu torchvision
Pillow
open_clip_torch
openai
fastapi
uvicorn